Python-projects/Rock-Paper-Scissor/requirements.txt
prettytable==2.1.0
mysql-connector-python==8.0.24